# Speed up leb128 write functions, add benchmarks