According to a few benchmarks I added for `leb128`, using a `scan`+`extend` combo instead of a push loop can improve the performance of write functions by **13-33%**:

before:
```
test leb128::tests::bench_write_u128  ... bench:       4,795 ns/iter (+/- 151)
test leb128::tests::bench_write_u16   ... bench:       4,795 ns/iter (+/- 122)
test leb128::tests::bench_write_u32   ... bench:       5,000 ns/iter (+/- 534)
test leb128::tests::bench_write_u64   ... bench:       5,013 ns/iter (+/- 235)
test leb128::tests::bench_write_usize ... bench:       5,011 ns/iter (+/- 119)
test leb128::tests::bench_write_i128  ... bench:       6,734 ns/iter (+/- 404)
```
after:
```
test leb128::tests::bench_write_u128  ... bench:       4,185 ns/iter (+/- 136)
test leb128::tests::bench_write_u16   ... bench:       3,697 ns/iter (+/- 516)
test leb128::tests::bench_write_u32   ... bench:       3,348 ns/iter (+/- 147)
test leb128::tests::bench_write_u64   ... bench:       3,367 ns/iter (+/- 180)
test leb128::tests::bench_write_usize ... bench:       3,381 ns/iter (+/- 393)
test leb128::tests::bench_write_i128  ... bench:       4,375 ns/iter (+/- 143)
```