@ljedrz 2018-08-06
If perf proves these gains I already have a similar improvement to `leb128::write_signed_leb128_to` ready :ok_hand:.