@ljedrz 2018-08-07
Hmm, the results aren't as awesome as indicated by the micro-benchmarks; are they invalid?

Even if the results of the change are not favourable it might be a good idea to keep the (possibly adjusted) benchmarks.