@ljedrz 2018-08-07
So it is entirely possible that some of the results are just noise? Hmm, any ideas on how the results of these changes can be measured?