@michaelwoerister 2018-08-07
> So it is entirely possible that some of the results are just noise?

I guess some of them are.

> Hmm, any ideas on how the results of these changes can be measured?

Since the amount of work a variable-length encoding depends on the actual numbers that one encodes, it's quite possible that the micro-benchmarks show a different picture than real-world test cases. I ran into this problem when trying to speed up leb128 earlier this year. To make optimizing the algorithm easier I created a testbed with more realistic test-data: https://github.com/michaelwoerister/encoding-bench (it's a bit of a mess and needs nightly to compile). I just tried to add your implementation but the function interface has changed (previously it had to support writing to anywhere in `out`).

> Perf will not show improvement if the compiler doesn't perform any LEB128 serialization in those test crates.

That's fine, the compiler does lots of leb128 encoding and decoding for crate metadata and incremental compilation, so one should definitely see a difference with a significantly faster algorithm.