@ljedrz 2018-08-07
@michaelwoerister thanks for the tip on `encoding-bench`! I'll use it to do some comparisons.