@ljedrz 2018-08-07
@michaelwoerister I [adjusted the `encoding-bench`](https://gist.github.com/ljedrz/b83287b446f9adfb7f7e536fc891c5f3) and got the following results:

before:
```
test write_leb128a_u16_query_cache   ... bench:          88 ns/iter (+/- 13) = 181 MB/s
test write_leb128a_u32_dep_graph     ... bench:   8,199,190 ns/iter (+/- 187,900) = 517 MB/s
test write_leb128a_u32_metadata      ... bench:     969,084 ns/iter (+/- 27,870) = 676 MB/s
test write_leb128a_u32_query_cache   ... bench:   2,801,147 ns/iter (+/- 72,513) = 831 MB/s
test write_leb128a_u64_dep_graph     ... bench:  17,231,910 ns/iter (+/- 315,007) = 312 MB/s
test write_leb128a_u64_metadata      ... bench:      96,983 ns/iter (+/- 4,074) = 364 MB/s
test write_leb128a_u64_query_cache   ... bench:   1,041,220 ns/iter (+/- 37,612) = 615 MB/s
test write_leb128a_usize_dep_graph   ... bench:     508,301 ns/iter (+/- 20,546) = 2652 MB/s
test write_leb128a_usize_metadata    ... bench:     746,033 ns/iter (+/- 39,989) = 2274 MB/s
test write_leb128a_usize_query_cache ... bench:   2,785,480 ns/iter (+/- 91,933) = 1562 MB/s
```
after:
```
test write_leb128b_u16_query_cache   ... bench:          82 ns/iter (+/- 3) = 195 MB/s
test write_leb128b_u32_dep_graph     ... bench:   5,281,614 ns/iter (+/- 177,362) = 802 MB/s
test write_leb128b_u32_metadata      ... bench:     641,292 ns/iter (+/- 34,882) = 1022 MB/s
test write_leb128b_u32_query_cache   ... bench:   2,167,868 ns/iter (+/- 70,851) = 1074 MB/s
test write_leb128b_u64_dep_graph     ... bench:  11,289,683 ns/iter (+/- 461,946) = 477 MB/s
test write_leb128b_u64_metadata      ... bench:      60,520 ns/iter (+/- 2,842) = 584 MB/s
test write_leb128b_u64_query_cache   ... bench:     676,636 ns/iter (+/- 28,241) = 946 MB/s
test write_leb128b_usize_dep_graph   ... bench:     517,218 ns/iter (+/- 21,342) = 2606 MB/s
test write_leb128b_usize_metadata    ... bench:     752,668 ns/iter (+/- 35,215) = 2254 MB/s
test write_leb128b_usize_query_cache ... bench:   2,936,698 ns/iter (+/- 114,159) = 1481 MB/s
```

It appears that the change is beneficial for everything besides `usize`, which regressed slightly. I was unable to compare `i128`, though (I only got bench results of 1ns) - perhaps you could tell what the issue is (it's included in my gist).