@michaelwoerister 2018-08-07
> I was unable to compare i128, though (I only got bench results of 1ns) - perhaps you could tell what the issue is

I assume that's because the dataset doesn't actually include any 128 bit numbers (we don't use them anywhere). 

I'm a bit surprised that the speedup for `u32`s doesn't translate to the compile times on perf.rlo. We are writing quite a few of those.