@michaelwoerister 2018-08-14
@ljedrz, I'm inclined to close this PR. The micro-benchmarks are a bit misleading and the new write implementation is harder to read without being a clear performance win. What do you say?